add `yAxisIndex` to trend line
