+++ frontend/src/metabase/visualizations/echarts/cartesian/option/utils.ts
@@ -1,0 +1,18 @@
+import type { CartesianChartModel, SeriesModel } from "../model/types";
+
+export function getSeriesYAxisIndex(
+  seriesModel: SeriesModel,
+  chartModel: CartesianChartModel,
+): number {
+  const hasSingleYAxis = chartModel.yAxisSplit.some(
+    yAxisKeys => yAxisKeys.length === 0,
+  );
+
+  if (hasSingleYAxis) {
+    return 0;
+  }
+
+  return chartModel.yAxisSplit.findIndex(yAxis =>
+    yAxis.includes(seriesModel.dataKey),
+  );
+}
